# laravel-project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Healthify API",
        "description": "API documentation for the Healthify project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/measurements/weights/chart/{range}": {
            "get": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Get weight chart data",
                "description": "Get weight chart data for the authenticated user.",
                "operationId": "8b12034d6e0aa3b247e53414beafc0ec",
                "parameters": [
                    {
                        "name": "range",
                        "in": "path",
                        "description": "Time range for the chart (days, weeks, months)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "days",
                            "enum": [
                                "days",
                                "weeks",
                                "months"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Weight chart data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "label": {
                                                                "type": "string",
                                                                "example": "2026-02-09"
                                                            },
                                                            "value": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 70.5
                                                            },
                                                            "start_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2026-02-09",
                                                                "nullable": true
                                                            },
                                                            "end_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2026-02-09",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "title": "Standard API Response",
                "description": "Standard wrapper for successful API responses",
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}